pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "blunder-butler"
version = "0.1.0"
description = "Chess.com game analysis and coaching report generator"
requires-python = ">=3.10"
dependencies = [
    "python-chess>=1.999",
    "stockfish>=3.28.0",
    "requests>=2.31.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "responses>=0.23",
    "ruff>=0.1.0",
]

[project.scripts]
blunder-butler = "blunder_butler.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
